[#3]: Change checks per seconf to float

+++ src/app/__main__.py
@@ -5,6 +5,6 @@ import sys
 from .app import init_dash_app
 
 if __name__ == "__main__":
-    check_per_s = int(sys.argv[1])
+    check_per_s = float(sys.argv[1])
     app = init_dash_app(1 / check_per_s)
     app.run_server(port=8050, debug=True)
